refactor(auth): split cloud broker internals
